pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
]
name = "tno-solar-nomad-plugin-base"
description = "Nomad example template"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Corne Frijters", email = "corne.frijters@tno.nl" },
]
maintainers = [
    { name = "Corne Frijters", email = "corne.frijters@tno.nl" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab>=1.3.0",
    "python-magic-bin; sys_platform == 'win32'",
    "numpy>=1.26.4",
]

[project.urls]
Repository = "https://github.com/TNO/tno-solar-nomad-plugin-base"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 200 # it's 2025, we have wide screens now
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
    "E501", # line too long 
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']
jv_parser = "tno_solar_nomad_plugin_base.jv:jv_parser_entry"
jv_schema = "tno_solar_nomad_plugin_base.jv:jv_schema_entry"
batch_parser = "tno_solar_nomad_plugin_base.batch:batch_parser_entry"
batch_schema = "tno_solar_nomad_plugin_base.batch:batch_schema_entry"

[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]
